Split out the mock streams
