+++ Cargo.toml
@@ -15,7 +15,6 @@ log = ">= 0.2"
 futures = "0.1"
 tokio-io = "0.1"
 tokio-core = "0.1.11"
-tokio-service = "0.1.0"
 tokio-reactor = "0.1.1"
 mio = "0.6"
